Change colours, implement VPN button

+++ lib/components/Date.jsx
@@ -1,7 +1,7 @@
 import { Calendar } from './Icons.jsx'
 
 const DateDisplay = () => {
-  const options = { weekday: 'long', month: 'long', day: 'numeric' }
+  const options = { weekday: 'short', month: 'short', day: 'numeric' }
   const now = new Date().toLocaleDateString('en-UK', options)
   return (
     <div className="date">
